+++ src/com/cosylab/epics/caj/impl/requests/EventAddRequest.java
@@ -106,14 +106,14 @@ public class EventAddRequest extends AbstractCARequest implements NotifyResponse
 		{
 		    requestMessage = ByteBuffer.allocate(CAConstants.CA_MESSAGE_HEADER_SIZE +  16);
 			requestMessage = insertCAHeader(transport, requestMessage,
-					(short)1, (short)16, (short)dataType, (short)dataCount,
+					(short)1, 16, (short)dataType, dataCount,
 					sid, subsid);
 		}
 		else 
 		{
 			requestMessage = ByteBuffer.allocate(CAConstants.CA_EXTENDED_MESSAGE_HEADER_SIZE + 16);
 		    requestMessage = insertCAHeader(transport, requestMessage,
-		            						(short)1, (short)0xFFFF, (short)dataType, (short)0,
+		            						(short)1, 0xFFFF, (short)dataType, 0,
 		            						sid, subsid);
 		    requestMessage.putInt(16);
 		    requestMessage.putInt(dataCount);
